split main package into subpackages
