+++ log/logwriter.go
@@ -1,4 +1,4 @@
-package main
+package log
 
 import (
 	"errors"
